Additional helpers for seccomp profile checks
